+++ tests/Integration/DSC_SqlSetup.config.ps1
@@ -52,7 +52,7 @@ else
 
                 # Database Engine properties.
                 DatabaseEngineNamedInstanceName         = 'DSCSQLTEST'
-                DatabaseEngineNamedInstanceFeatures     = 'SQLENGINE,AS,CONN,BC,SDK'
+                DatabaseEngineNamedInstanceFeatures     = 'SQLENGINE,REPLICATION,AS,CONN,BC,SDK'
                 AnalysisServicesMultiServerMode         = 'MULTIDIMENSIONAL'
 
                 <#
@@ -74,7 +74,7 @@ else
                     evaluate the result in the test.
                 #>
                 DatabaseEngineDefaultInstanceName       = 'MSSQLSERVER'
-                DatabaseEngineDefaultInstanceFeatures   = 'SQLENGINE,CONN,BC,SDK'
+                DatabaseEngineDefaultInstanceFeatures   = 'SQLENGINE,REPLICATION,CONN,BC,SDK'
 
                 # General SqlSetup properties
                 Collation                               = 'Finnish_Swedish_CI_AS'